requirements.txt
openai-whisper
PyAudio
pyperclip
PySimpleGUI

